pyproject.toml
[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[project]
name = "fantasy-recipe"
version = "2.0.0"
description = "Fantasy recipes"
authors = []
dependencies = [
    "alembic>=1.7.7",
    "anyio[trio]>=3.5.0",
    "asyncpg>=0.25.0",
    "fastapi>=0.75.0",
    "Jinja2>=3.1.1",
    "passlib[argon2]>=1.7.4",
    "pydantic>=1.9.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.5",
    "SQLAlchemy[asyncio]>=1.4.35",
    "uvicorn[standard]>=0.17.6"
]
requires-python = ">=3.9"
license = {text = "MIT"}

[project.urls]
homepage = ""

[project.optional-dependencies]
[tool]
[tool.pdm.scripts]
server = "uvicorn main:app --host 0.0.0.0 --port $PORT --reload"
test = "coverage run -m pytest tests/ --junitxml=junit.xml"
post_test = "coverage report -m"

[tool.pdm.dev-dependencies]
dev = [
    "black>=22.3.0",
    "coverage[toml]>=6.3",
    "mypy>=0.942",
    "pytest>=7.1.1",
    "requests>=2.27.1"
]

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
ignore_missing_imports = true
warn_no_return = false
warn_return_any = false
exclude = [
    "^__pypackages__/"
]
[tool.pytest.ini_options]
markers = [
    "webtest: marks tests as slow (deselect with '-m \"not slow\"')",
]
[tool.coverage.run]
branch = true
omit = [
    "tests/*",
    "__pypackages__/*"
]

[tool.coverage.report]
skip_empty = true
show_missing = true
